set localstack docker image to specific version, imporved test
