+++ worker/src/test/java/gov/cms/ab2d/worker/service/JobUpdateListenerServiceITest.java
@@ -2,6 +2,7 @@ package gov.cms.ab2d.worker.service;
 
 import com.amazonaws.services.sns.AmazonSNS;
 import com.amazonaws.services.sqs.AmazonSQS;
+import com.amazonaws.services.sqs.model.QueueDoesNotExistException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gov.cms.ab2d.common.util.AB2DLocalstackContainer;
@@ -11,6 +12,7 @@ import gov.cms.ab2d.worker.dto.SNSMessage;
 import gov.cms.ab2d.worker.processor.JobProgressService;
 import gov.cms.ab2d.worker.processor.JobProgressUpdateService;
 import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,10 +21,13 @@ import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
+import java.util.Random;
+import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 import static gov.cms.ab2d.worker.processor.JobMeasure.FAILURE_THRESHHOLD;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.testcontainers.shaded.org.awaitility.Awaitility.await;
 
 @Testcontainers
@@ -54,20 +59,35 @@ public class JobUpdateListenerServiceITest {
     @Container
     private static final AB2DLocalstackContainer localstackContainer = new AB2DLocalstackContainer();
 
+
+    @Test
+    @DisplayName("SQS queue delete")
+    void sqsCleanup() {
+        JobUpdateListenerServiceImpl listener = new JobUpdateListenerServiceImpl(amazonSqs, amazonSns, mapper, jobProgressUpdateService);
+        ReflectionTestUtils.invokeMethod(listener, "initiate");
+        ReflectionTestUtils.invokeMethod(listener, "cleanup");
+        assertThrows(QueueDoesNotExistException.class, () -> {
+            String url = String.valueOf(ReflectionTestUtils.getField(jobUpdateListenerService, "queueUrl"));
+            amazonSqs.getQueueUrl(url);
+        });
+    }
+
     @Test
     void pollSqs() throws JsonProcessingException {
+        final int oldThreshold = new Random().nextInt();
+        final int newThreshold = new Random().nextInt();
         String url = String.valueOf(ReflectionTestUtils.getField(jobUpdateListenerService, "queueUrl"));
         jobProgressUpdateService.initJob("test");
-        jobProgressUpdateService.addMeasure("test", FAILURE_THRESHHOLD, 1);
+        jobProgressUpdateService.addMeasure("test", FAILURE_THRESHHOLD, oldThreshold);
         SNSMessage snsMessage = new SNSMessage();
         snsMessage.setSubject("test");
         snsMessage.setMessage(mapper.writeValueAsString(JobUpdate.builder()
                 .measure(String.valueOf(FAILURE_THRESHHOLD))
-                .value(10)
+                .value(newThreshold)
                 .build()));
         amazonSqs.sendMessage(url, mapper.writeValueAsString(snsMessage));
         await().atMost(10, TimeUnit.SECONDS).until(() ->
-                jobProgressService.getStatus("test").getFailureThreshold() == 10);
+                jobProgressService.getStatus("test").getFailureThreshold() != oldThreshold);
     }
 
     @Test
@@ -93,13 +113,20 @@ public class JobUpdateListenerServiceITest {
         );
     }
 
+
     @Test
-    void sqsCleanup() {
-        JobUpdateListenerServiceImpl listener = new JobUpdateListenerServiceImpl(amazonSqs, amazonSns, mapper, jobProgressUpdateService);
-        ReflectionTestUtils.invokeMethod(listener, "initiate");
-        assertDoesNotThrow(() ->
-                ReflectionTestUtils.invokeMethod(listener, "cleanup")
+    @DisplayName("Update job progress by calling SNS handler method")
+    void jobUpdateDirectly() {
+        final int oldThreshold = new Random().nextInt();
+        final int newThreshold = new Random().nextInt();
+        String uuid = UUID.randomUUID().toString();
+        log.info("sending directly uuid: {} old: {} new: {}", uuid, oldThreshold, newThreshold);
+        jobProgressUpdateService.initJob(uuid);
+        jobProgressUpdateService.addMeasure(uuid, FAILURE_THRESHHOLD, newThreshold);
+        await().atMost(10, TimeUnit.SECONDS).until(() ->
+                newThreshold == jobProgressService.getStatus(uuid).getFailureThreshold()
         );
     }
+
 
 }
